requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
jinja2==3.1.2
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
statsmodels==0.14.0
joblib==1.3.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pulp==2.7.0
apscheduler==3.10.4